twitoff/requirements.txt
alembic==1.0.8
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
basilica==0.2.5
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Mako==1.0.7
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.2
oauthlib==3.0.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
ptyprocess==0.6.0
Pygments==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
python-decouple==3.1
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
SQLAlchemy==1.2.18
terminado==0.8.1
testpath==0.4.2
tornado==6.0.1
traitlets==4.3.2
tweepy==3.7.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2